Refactor AnnotationValueExpressionUtil
Merge the logic of ModifyAnnotationAttributes into AnnotationValueExpressionUtil
Test case modification
Update reference of ModifyAnnotationAttributes
